Create README.md
You will complete the microcode for seven brand-new LC-3 instructions. For each instruction you should fill out the corresponding rows in the tl02microcode.xlsx sheet. You need to implement INC, LDPC, LJSR, MULT8, LDAR, STPCI, and CCH, as described below. Note, when talking about registers, DR, SR1, and SR2, they can either refer to the 3 bits that indicate WHICH register (we have 8 total), or refer to the 16 bit value contained within a register. In the 16 bit instruction layouts, SR1 would refer to the 3 bits (as seen by the bit range 9:11 for example). In a line like SR1 <= SR1 + sext(imm9), this in plain English means: SR1 takes the value of SR1’s previous value plus a 9 bit offset number.

    • Hints

        ◦ Fill in all unused control signals with 0.

        ◦ If you are lost, try seeing how each instruction compares to those that you completed in the HW4, and transform your HW4 instructions into the ones listed above. This does not work for all instructions, but will work for a few.

        ◦ Do not forget to set the condition codes for the instructions that require it!


    • Instructions

4.1	Writing the Microcode

Use the provided tl02microcode.xlsx file and fill in the control signals for each instruction.Do not edit any of the values that have already been provided.

Also, please refer to Section 8.1 if you need to know which MUX selector value corresponds to which input, or reference the LC3.sim file. Do NOT use the LC3 reference sheet to guess the order of pins.